+++ modules/math/app/services/equation/EquationGenerator.java
@@ -1,16 +1,11 @@
 package services.equation;
 
-import java.math.BigInteger;
+import java.math.BigDecimal;
 import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.List;
 import java.util.Random;
-
 import beans.math.MathBean;
-import models.math.MathQuestion;
 import services.categories.Categorizer;
-import services.categories.detector.CategoryDetector;
-import services.categories.detector.PerimeterDetectorSquare;
 
 public class EquationGenerator {
 	private Random random = new Random();
@@ -44,20 +39,17 @@ public class EquationGenerator {
 
 		boolean firstIteration = true;
 		for (int i = 0; i < iterations; i++) {
-			BigInteger number = getRandomNumber();
+			BigDecimal number = getRandomNumber();
 			basicOperator = getRandomOperation();
 			
 			if (firstIteration) {
 				mathBean.getIntegers().add(number);
 				number = getRandomNumber();
-			
-				
 				firstIteration = false;
 			}
 			if (basicOperator.equals(Operators.DIVIDE)) {
-				number = BigInteger.valueOf(random.nextInt(2) + 2); // only
-																	// divide by
-																	// 2 or 3
+				// only divide by 2 or 3
+				number = BigDecimal.valueOf(random.nextInt(2) + 2);
 			}
 			
 			mathBean.getIntegers().add(number);
@@ -80,8 +72,8 @@ public class EquationGenerator {
 		return Operators.WEIGHTED_OPERATORS.get(random.nextInt(Operators.WEIGHTED_OPERATORS.size()));
 	}
 
-	private BigInteger getRandomNumber() {
-		return BigInteger.valueOf(random.nextInt(9) + 1);
+	private BigDecimal getRandomNumber() {
+		return BigDecimal.valueOf(random.nextInt(9) + 1);
 	}
 
 }
